Updated Tags, Updated MechDef

+++ Code/Utils/C#/BattleEngineJsonCreation/BattleEngineJsonCreation/JSONClass/Mechs/MechDef.cs
@@ -21,7 +21,7 @@ namespace BattleEngineJsonCreation
         public long? SimGameMechPartCost { get; set; }
 
         [JsonProperty("Version", NullValueHandling = NullValueHandling.Ignore)]
-        public double? Version { get; set; }
+        public int? Version { get; set; }
 
         [JsonProperty("Locations", NullValueHandling = NullValueHandling.Ignore)]
         public List<MechDefLocation> Locations { get; set; }
